comments fixes - moved the methods into the parent class so they can be used in other tests as well
